day 10 in C#
I need to do this things in a less distracting environment, or start listening to music or something.
This is also the (2nd? 3rd?) time that I have to use an int64 - obvious in retrospect,
but it makes me wonder if the authors use a language where numbers default to 64 bits (golang)
or something with infinite precision (lisp family, python).