day 14 in common lisp
Late because real life happened, and THEN my internet died yesterday and I didn't have the common lisp hyperthing offline...
In common lisp because that's a language that was still missing - now what? lua? stop this "use a random different language" thing?

Anyway there's a ton of output in part2 because I was looking for the most asinine bug ever (commented out to let posterity know I can be so stupid);
it works in clisp, sbcl for whatever reason does strange things, and while I love sbcl's speed & diagnostics I'm not debugging that while I'm late on the schedule, mkay?